Driver Mock Test screen implementation completed

+++ DriveSense/DriverSenseModule/Datalayer/Enitiy/AxisValueModel.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct AxisValueModel:Codable {
-    var x: Double?
-    var y: Double?
-    var z: Double?
+    var x: Float?
+    var y: Float?
+    var z: Float?
 }
